refactor(navbar): move profile button to internal components
