+++ projects/client/src/lib/sections/navbar/SideNavbar.svelte
@@ -13,8 +13,8 @@
   import { UrlBuilder } from "$lib/utils/url/UrlBuilder";
   import BetaBadge from "./components/BetaBadge.svelte";
   import FilterButton from "./components/filter/FilterButton.svelte";
+  import ProfileButton from "./components/ProfileButton.svelte";
   import TraktLogo from "./components/TraktLogo.svelte";
-  import ProfileButton from "./ProfileButton.svelte";
 
   const isTV = $derived(browser && getDeviceType(navigator.userAgent) === "tv");
 </script>
